requirements.txt
pandas
python-dateutil
matplotlib
seaborn
python-pptx
Pillow